backend/requirements.txt
# ==================== FRAMEWORK PRINCIPAL ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== BASE DE DATOS ====================
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # Para PostgreSQL
redis==5.0.1

# ==================== AUTENTICACIÓN Y SEGURIDAD ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ==================== CONFIGURACIÓN Y ENTORNO ====================
python-dotenv==1.0.0
pydantic==2.5.0

# ==================== MACHINE LEARNING ====================
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1

# ==================== SERVICIOS EXTERNOS ====================
httpx==0.25.2
aiohttp==3.9.1
openai==1.3.7

# ==================== LOGGING Y MONITOREO ====================
structlog==23.2.0
prometheus-client==0.19.0

# ==================== VALIDACIÓN Y UTILIDADES ====================
email-validator==2.1.0
python-dateutil==2.8.2

# ==================== TESTING ====================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# ==================== DESARROLLO ====================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ==================== PRODUCCIÓN ====================
gunicorn==21.2.0
sentry-sdk[fastapi]==1.38.0

# ==================== CACHÉ Y OPTIMIZACIÓN ====================
cachetools==5.3.2
aioredis==2.0.1

# ==================== NOTIFICACIONES ====================
celery==5.3.4
flower==2.0.1

# ==================== DOCUMENTACIÓN ====================
mkdocs==1.5.3
mkdocs-material==9.4.8

# ==================== DEPENDENCIAS ESPECÍFICAS ====================
# Para análisis de texto
textblob==0.17.1
spacy==3.7.2

# Para procesamiento de datos
python-json-logger==2.0.7

# Para rate limiting
slowapi==0.1.9

# Para validación de esquemas
marshmallow==3.20.1

# Para manejo de archivos
python-magic==0.4.27
Pillow==10.1.0

# Para compresión
gzip==1.0.0

# Para encriptación adicional
cryptography==41.0.8

# Para manejo de fechas avanzado
arrow==1.3.0

# Para validación de URLs
validators==0.22.0

# Para manejo de configuraciones
dynaconf==3.2.4

# Para métricas de rendimiento
py-spy==0.3.14

# Para profiling
memory-profiler==0.61.0

# Para testing de integración
testcontainers==3.7.1

# Para documentación automática
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
